move debugMessages query to admin gql file
